Refactor to remove redundant monthlyContainer
Logic moved to DateRows
