+++ frontend/src/components/Calendar/DateRows.tsx
@@ -2,27 +2,49 @@ import React from 'react'
 import { ClickableDateContainer } from './ClickableDateContainer'
 import { EmptyDateContainer } from './EmptyDateContainer'
 
-export const DateRows = (props: {
-  firstDate: number
-  assignedMonth: number
-  lastDateOfThisMonth: number
-  bufferDates: number[]
-  facilityId: number
-}) => {
-  for (let i = props.firstDate; i < props.lastDateOfThisMonth + 1; i++) {
-    props.bufferDates.push(i)
+export const DateRows = (props: { currentDate: Date; nthMonth: number; facilityId: number }) => {
+  const firstDateOfThisMonth = new Date(
+    props.currentDate.getFullYear(),
+    props.currentDate.getMonth() + props.nthMonth,
+    1,
+  ).getDate()
+  const assignedMonth = props.currentDate.getMonth() + props.nthMonth
+  const firstDayOfThisMonth = new Date(
+    props.currentDate.getFullYear(),
+    props.currentDate.getMonth() + props.nthMonth,
+    1,
+  ).getDay()
+  const lastDateOfThisMonth = new Date(
+    props.currentDate.getFullYear(),
+    props.currentDate.getMonth() + 1 + props.nthMonth,
+    0,
+  ).getDate()
+  const bufferDates: number[] = []
+
+  if (firstDayOfThisMonth === 0) {
+    for (let i = 1; i < 7; i++) {
+      bufferDates.push(0)
+    }
+  } else {
+    for (let i = 1; i < firstDayOfThisMonth; i++) {
+      bufferDates.push(0)
+    }
+  }
+
+  for (let i = firstDateOfThisMonth; i < lastDateOfThisMonth + 1; i++) {
+    bufferDates.push(i)
   }
 
   return (
     <>
-      {props.bufferDates.map((day) => {
+      {bufferDates.map((day) => {
         return day === 0 ? (
           <EmptyDateContainer />
         ) : (
           <ClickableDateContainer
-            key={day + props.assignedMonth * 100}
+            key={day + assignedMonth * 100}
             date={day}
-            assignedMonth={props.assignedMonth}
+            assignedMonth={assignedMonth}
             facilityId={props.facilityId}
           />
         )
